requirements.txt
streamlit
joblib
scikit-learn
python-docx
reportlab
